Make PowerShell script more handy for dev
It mostly equivalent in capability to makefile, though `release` is
orthogonal to whether building, checking, or running, and can be applied to
any of them.

I am a PowerShell noob so there are probably some issues with this, but
it appears to work.

Notably, PowerShell isn't really optimised for logic based on "exit statuses"
and instead models success/failure quite reasonably around which output
streams a function/command/etc writes to. This isn't really useful when we
are predominantly shimming over binaries modelled in the Unix world, so this
script appears pretty clumsy.